Move DTOs to model package
