+++ extensions/dataspace-authority-spi/src/main/java/org/eclipse/dataspaceconnector/registration/authority/dto/ParticipantDto.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.dataspaceconnector.registration.dto;
+package org.eclipse.dataspaceconnector.registration.authority.dto;
 
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
