+++ extensions/registration-service/src/main/java/org/eclipse/dataspaceconnector/registration/api/RegistrationService.java
@@ -16,7 +16,7 @@ package org.eclipse.dataspaceconnector.registration.api;
 
 import org.eclipse.dataspaceconnector.api.transformer.DtoTransformerRegistry;
 import org.eclipse.dataspaceconnector.registration.authority.model.Participant;
-import org.eclipse.dataspaceconnector.registration.dto.ParticipantDto;
+import org.eclipse.dataspaceconnector.registration.authority.dto.ParticipantDto;
 import org.eclipse.dataspaceconnector.registration.store.spi.ParticipantStore;
 import org.eclipse.dataspaceconnector.spi.exception.ObjectNotFoundException;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
