+++ extensions/registration-service/src/main/java/org/eclipse/dataspaceconnector/registration/transform/ParticipantToParticipantDtoTransformer.java
@@ -17,8 +17,8 @@ package org.eclipse.dataspaceconnector.registration.transform;
 import org.eclipse.dataspaceconnector.api.transformer.DtoTransformer;
 import org.eclipse.dataspaceconnector.registration.authority.model.Participant;
 import org.eclipse.dataspaceconnector.registration.authority.model.ParticipantStatus;
-import org.eclipse.dataspaceconnector.registration.dto.ParticipantDto;
-import org.eclipse.dataspaceconnector.registration.dto.ParticipantStatusDto;
+import org.eclipse.dataspaceconnector.registration.authority.dto.ParticipantDto;
+import org.eclipse.dataspaceconnector.registration.authority.dto.ParticipantStatusDto;
 import org.eclipse.dataspaceconnector.spi.EdcException;
 import org.eclipse.dataspaceconnector.spi.transformer.TransformerContext;
 import org.jetbrains.annotations.NotNull;
